Extract logic from AbstractDocumentationCommentCommandHandler into a separate service
