+++ src/EditorFeatures/Core/Implementation/DocumentationComments/IDocumentationCommentSnippetService.cs
@@ -1,0 +1,44 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Threading;
+using Microsoft.CodeAnalysis.Host;
+using Microsoft.CodeAnalysis.Options;
+using Microsoft.CodeAnalysis.Text;
+
+namespace Microsoft.CodeAnalysis.Editor
+{
+    internal interface IDocumentationCommentSnippetService : ILanguageService
+    {
+        DocumentationCommentSnippet GetDocumentationCommentSnippetOnCharacterTyped(
+            SyntaxTree syntaxTree,
+            SourceText text,
+            int position,
+            DocumentOptionSet options,
+            CancellationToken cancellationToken);
+
+        DocumentationCommentSnippet GetDocumentationCommentSnippetOnCommandInvoke(
+            SyntaxTree syntaxTree,
+            SourceText text,
+            int position,
+            DocumentOptionSet options,
+            CancellationToken cancellationToken);
+
+        DocumentationCommentSnippet GetDocumentationCommentSnippetOnEnterTyped(
+            SyntaxTree syntaxTree,
+            SourceText text,
+            int position,
+            DocumentOptionSet options,
+            string exteriorTriviaText,
+            CancellationToken cancellationToken);
+
+        DocumentationCommentSnippet GetDocumentationCommentSnippetFromPreviousLine(
+            DocumentOptionSet options,
+            string exteriorTriviaText,
+            TextLine currentLine,
+            TextLine previousLine);
+
+        bool IsValidTargetMember(SyntaxTree syntaxTree, SourceText text, int caretPosition, CancellationToken cancellationToken);
+    }
+}
